refactor(vscode): extract webview-html with a parameterized action button
Both the sidebar and the forthcoming editor panel render identical HTML, so
renderHtml/escapeHtml/makeNonce move to their own vscode-free module. The
status pane's hardcoded 'Retry' -> claudeUsage.open button becomes a
caller-supplied WebviewAction, which the editor panel needs to point its own
retry at claudeUsage.openInEditor.
